вынес вспомогательные функции в utls.js. Поправил выбор активного фильтра в main.js
